+++ NINA/ViewModel/FramingAssistant/FramingAssistantVM.cs
@@ -553,6 +553,7 @@ namespace NINA.ViewModel.FramingAssistant {
             set {
                 if (RectangleCalculated) {
                     Rectangle.TotalRotation = value;
+                    profileService.ActiveProfile.FramingAssistantSettings.LastRotationAngle = value;
                     RaisePropertyChanged();
                     RaisePropertyChanged(nameof(RectangleRotation));
                     DragMove(new DragResult() { Delta = new Vector() });
@@ -1136,8 +1137,8 @@ namespace NINA.ViewModel.FramingAssistant {
 
                         await SkyMapAnnotator.Initialize(skySurveyImage.Coordinates, AstroUtil.ArcminToDegree(skySurveyImage.FoVHeight), ImageParameter.Image.PixelWidth, ImageParameter.Image.PixelHeight, ImageParameter.Rotation, Cache, _loadImageSource.Token);
                         SkyMapAnnotator.DynamicFoV = FramingAssistantSource == SkySurveySource.SKYATLAS;
-
                         CalculateRectangle(SkyMapAnnotator.ViewportFoV);
+                        RectangleTotalRotation = profileService.ActiveProfile.FramingAssistantSettings.LastRotationAngle;
                     }
                 } catch (OperationCanceledException) {
                     Logger.Info("Loading image for framing has been cancelled");